module1_3requirements.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aioredis==2.0.1
aiosignal==1.2.0
angel-cd==1.0.3
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bimlpa==0.1.2
bleach==5.0.0
blessings==1.7
cachetools==5.1.0
cdlib==0.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
chinese-whispers==0.8.0
click==8.1.3
colorama==0.4.4
colorful==0.5.4
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
demon==2.0.6
dynetx==0.3.1
entrypoints==0.4
eva-lcd==0.1.1
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.0
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
gensim==4.2.0
google==3.0.0
google-api-core==2.7.3
google-auth==2.6.6
googleapis-common-protos==1.56.1
gpustat==0.6.0
grpcio==1.46.1
idna==3.3
igraph==0.9.10
importlib-resources==5.7.1
infomap==2.3.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab==2.2.2
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
karateclub==1.2.3
kiwisolver==1.4.2
leidenalg==0.8.10
llvmlite==0.38.0
markov-clustering==0.0.6.dev0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
multidict==6.0.2
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
nf1==0.0.4
notebook==6.1.0
numba==0.55.1
numpy==1.20.3
nvidia-ml-py3==7.352.0
opencensus==0.9.0
opencensus-context==0.1.2
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pip==20.0.2
pkg-resources==0.0.0
pooch==1.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
PuLP==2.6.0
pure-eval==0.2.2
py-spy==0.3.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclustering==0.10.1.2
pycparser==2.21
Pygments==2.12.0
PyGSP==0.5.1
pynndescent==0.5.7
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-igraph==0.9.10
python-Levenshtein==0.12.2
python-louvain==0.16
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
ray==1.0.1.post1
redis==3.4.1
requests==2.27.1
rsa==4.8
scikit-learn==1.1.0
scipy==1.8
seaborn==0.11.1
Send2Trash==1.8.0
setuptools==44.0.0
six==1.16.0
smart-open==6.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
terminado==0.15.0
texttable==1.6.4
threadpoolctl==3.1.0
thresholdclustering==1.1
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typing-extensions==4.2.0
umap-learn==0.5.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
wurlitzer==3.0.2
yarl==1.7.2
zipp==3.8.0
